gpt.gpt: More modular approach to scanning GPT
At start I thought I will need to implement only one PReg
parser/interpreter and that's it. At the moment it is demanded all
parser to be implemented, all GPT settings covered and ADMX files to
be extended to work with Linux (ALT) specific cases.

Now I have more expertise in the domain so this PR doint simple thing
despite lots of code:

* There is "parser" for each file type in corresponding module.
* There is also "merger" for each object parsed from file in
  corresponding module.

We're doing this:

* Determine file type by its name (stupid and working!)
* Call the corresponding parser to get objects
* Call the corresponding merger for objects